add a test which I believe shows a BUG (but I could be mistaken)
the file named:

tools.ietf.org_doc_groff_meintro.ps.gz.2012-06-25.v1.21.0.me.ps.sanitized.SHOULD

... is a PS file that I downloaded from ietf.org, which was timestamped
with the year 2012.

if you run the current codebase groff binary on:

tools.ietf.org_doc_groff_meintro.ps.gz.2012-06-25.v1.21.0.me

... you do not get the same output as in 2012.

If you further convert PS to PDF, you can see that the OLDER PS file
generates something much nicer-looking on page 13.

Page 13 is the most obvious change (and in a direction that is
more clearly undesirable).

There is also a change on page 1, in the last paragraph of the page.
In the older version, the word 'Arguments' is not split across a
line break, and in the newer version it is.